reorganized preprocessing module
